webhook_gcp/requirements.txt
requests
pandas
google-cloud-storage
google-cloud-tasks
scikit-learn==1.3.2
xgboost==2.0.3
joblib